dns-scanner/code/requirements.txt
PyYAML
dnspython
